+++ uCOS-II/os_cfg.h
@@ -57,15 +57,12 @@
 #define OS_LOWEST_PRIO    RT_THREAD_PRIORITY_MAX /* 只读 Defines the lowest priority that can be assigned...*/
                                        /* ... MUST NEVER be higher than 254!                           */
 
-#define OS_MAX_FLAGS              5u   /* Max. number of Event Flag Groups    in your application      */
 #define OS_MAX_MEM_PART           5u   /* Max. number of memory partitions                             */
-#define OS_MAX_QS                 4u   /* Max. number of queue control blocks in your application      */
 #define OS_MAX_TASKS             20u   /* Max. number of tasks in your application, MUST be >= 2       */
 
 #define OS_SCHED_LOCK_EN          1u   /* Include code for OSSchedLock() and OSSchedUnlock()           */
 
-#define OS_TICK_STEP_EN           1u   /* Enable tick stepping feature for uC/OS-View                  */
-#define OS_TICKS_PER_SEC  RT_TICK_PER_SECOND   /* 只读 Set the number of ticks in one second           */
+#define OS_TICKS_PER_SEC  RT_TICK_PER_SECOND /* 只读 Set the number of ticks in one second             */
 
 #define OS_TLS_TBL_SIZE           0u   /* Size of Thread-Local Storage Table                           */
 
@@ -90,16 +87,16 @@
 
                                        /* ----------------------- EVENT FLAGS ------------------------ */
 #ifdef RT_USING_EVENT                  /* 是否开启由RT-Thread接管                                      */
-#define OS_FLAG_EN                1u   /* Enable (1) or Disable (0) code generation for EVENT FLAGS    */
+#define OS_FLAG_EN                1u   /* 读写 Enable (1) or Disable (0) code generation for EVENT FLAGS*/
 #else
-#define OS_FLAG_EN                0u   /* Enable (1) or Disable (0) code generation for EVENT FLAGS    */
+#define OS_FLAG_EN                0u   /* 只读 Enable (1) or Disable (0) code generation for EVENT FLAGS*/
 #endif
-#define OS_FLAG_ACCEPT_EN         1u   /*     Include code for OSFlagAccept()                          */
-#define OS_FLAG_DEL_EN            1u   /*     Include code for OSFlagDel()                             */
-#define OS_FLAG_NAME_EN           1u   /*     Enable names for event flag group                        */
-#define OS_FLAG_QUERY_EN          1u   /*     Include code for OSFlagQuery()                           */
-#define OS_FLAG_WAIT_CLR_EN       1u   /* Include code for Wait on Clear EVENT FLAGS                   */
-#define OS_FLAGS_NBITS           32u   /* Size in #bits of OS_FLAGS data type (8, 16 or 32)            */
+#define OS_FLAG_ACCEPT_EN         0u   /*     Include code for OSFlagAccept()                          */
+#define OS_FLAG_DEL_EN            0u   /*     Include code for OSFlagDel()                             */
+#define OS_FLAG_NAME_EN           0u   /*     Enable names for event flag group                        */
+#define OS_FLAG_QUERY_EN          0u   /*     Include code for OSFlagQuery()                           */
+#define OS_FLAG_WAIT_CLR_EN       0u   /* Include code for Wait on Clear EVENT FLAGS                   */
+#define OS_FLAGS_NBITS           16u   /* Size in #bits of OS_FLAGS data type (8, 16 or 32)            */
 
 
                                        /* --------------------- MEMORY MANAGEMENT -------------------- */
@@ -110,9 +107,9 @@
 
                                        /* ---------------- MUTUAL EXCLUSION SEMAPHORES --------------- */
 #ifdef RT_USING_MUTEX                  /* 是否开启由RT-Thread接管                                      */
-#define OS_MUTEX_EN               1u   /* Enable (1) or Disable (0) code generation for MUTEX          */
+#define OS_MUTEX_EN               1u   /* 读写 Enable (1) or Disable (0) code generation for MUTEX     */
 #else
-#define OS_MUTEX_EN               0u   /* Enable (1) or Disable (0) code generation for MUTEX          */
+#define OS_MUTEX_EN               0u   /* 只读 Enable (1) or Disable (0) code generation for MUTEX     */
 #endif
 #define OS_MUTEX_ACCEPT_EN        1u   /*     Include code for OSMutexAccept()                         */
 #define OS_MUTEX_DEL_EN           1u   /*     Include code for OSMutexDel()                            */
@@ -121,9 +118,9 @@
 
                                        /* ---------------------- MESSAGE QUEUES ---------------------- */
 #ifdef RT_USING_MESSAGEQUEUE           /* 是否开启由RT-Thread接管                                      */
-#define OS_Q_EN                   1u   /* Enable (1) or Disable (0) code generation for QUEUES         */
+#define OS_Q_EN                   1u   /* 读写 Enable (1) or Disable (0) code generation for QUEUES    */
 #else
-#define OS_Q_EN                   0u   /* Enable (1) or Disable (0) code generation for QUEUES         */
+#define OS_Q_EN                   0u   /* 只读 Enable (1) or Disable (0) code generation for QUEUES    */
 #endif
 #define OS_Q_ACCEPT_EN            1u   /*     Include code for OSQAccept()                             */
 #define OS_Q_DEL_EN               1u   /*     Include code for OSQDel()                                */
@@ -136,7 +133,7 @@
 
 
                                        /* -------------------- MESSAGE MAILBOXES --------------------- */
-#define OS_MBOX_EN           OS_Q_EN   /* Enable (1) or Disable (0) code generation for MAILBOXES      */
+#define OS_MBOX_EN           OS_Q_EN   /* 只读 Enable (1) or Disable (0) code generation for MAILBOXES */
 #define OS_MBOX_ACCEPT_EN         1u   /*     Include code for OSMboxAccept()                          */
 #define OS_MBOX_DEL_EN            1u   /*     Include code for OSMboxDel()                             */
 #define OS_MBOX_PEND_ABORT_EN     1u   /*     Include code for OSMboxPendAbort()                       */
@@ -147,9 +144,9 @@
 
                                        /* ------------------------ SEMAPHORES ------------------------ */
 #ifdef RT_USING_SEMAPHORE              /* 是否开启由RT-Thread接管                                      */
-#define OS_SEM_EN                 1u   /* Enable (1) or Disable (0) code generation for SEMAPHORES     */
+#define OS_SEM_EN                 1u   /* 读写 Enable (1) or Disable (0) code generation for SEMAPHORES*/
 #else
-#define OS_SEM_EN                 0u   /* Enable (1) or Disable (0) code generation for SEMAPHORES     */
+#define OS_SEM_EN                 0u   /* 只读 Enable (1) or Disable (0) code generation for SEMAPHORES*/
 #endif
 #define OS_SEM_ACCEPT_EN          1u   /*    Include code for OSSemAccept()                            */
 #define OS_SEM_DEL_EN             1u   /*    Include code for OSSemDel()                               */
@@ -166,9 +163,9 @@
 
                                        /* --------------------- TIMER MANAGEMENT --------------------- */
 #ifdef RT_USING_TIMER_SOFT             /* 是否开启由RT-Thread接管                                      */
-#define OS_TMR_EN                 1u   /* Enable (1) or Disable (0) code generation for TIMERS         */
+#define OS_TMR_EN                 1u   /* 读写 Enable (1) or Disable (0) code generation for TIMERS    */
 #else
-#define OS_TMR_EN                 0u   /* Enable (1) or Disable (0) code generation for TIMERS         */
+#define OS_TMR_EN                 0u   /* 只读 Enable (1) or Disable (0) code generation for TIMERS    */
 #endif
 #define OS_TMR_CFG_NAME_EN        1u   /*     Determine timer names                                    */
 #define OS_TMR_CFG_TICKS_PER_SEC 10u   /*     Rate at which timer management task runs (Hz)            */